+++ apps/bam.iobio/src/util.ts
@@ -19,7 +19,7 @@
  *
  */
 
-import type { BamContext, BamKey } from 'components/src/constants';
+import type { BamContext, BamKey } from 'iobio-react-components/src/constants';
 
 export const iobioURL = 'https://s3.amazonaws.com/iobio/NA12878/NA12878.autsome.bam';
 
@@ -37,27 +37,6 @@ export const defaultBamContext = {
 	mapq_hist: true,
 	baseq_hist: true,
 } as const satisfies BamContext;
-
-export const histogramKeys = [
-	'coverage_hist',
-	'frag_hist',
-	'length_hist',
-	'mapq_hist',
-	'baseq_hist',
-] as const satisfies Array<BamKey>;
-
-export type BamHistogramKey = (typeof histogramKeys)[number];
-
-export const percentKeys = [
-	'mapped_reads',
-	'forward_strands',
-	'proper_pairs',
-	'singletons',
-	'both_mates_mapped',
-	'duplicates',
-] as const satisfies Array<BamKey>;
-
-export type BamPercentKey = (typeof percentKeys)[number];
 
 const ignoreOutlierKeys = ['frag_hist', 'length_hist'] as const satisfies Array<BamKey>;
