+++ packages/iobio-react-components/src/coverage_depth.tsx
@@ -24,3 +24,5 @@ function IobioCoverageDepth({ label }: { label?: string }) {
 }
 
 export default IobioCoverageDepth;
+
+export type IobioCoverageDepthType = typeof IobioCoverageDepth;
