+++ packages/iobio-react-components/src/data_broker.tsx
@@ -24,3 +24,5 @@ function IobioDataBroker({ alignmentUrl }: { alignmentUrl: string }) {
 }
 
 export default IobioDataBroker;
+
+export type IobioDataBrokerType = typeof IobioDataBroker;
